Added different message for empty required fields and a different dialog.  This will allow the user to intentionally skip an entry.  All fields will be required when the final validation is performed.  This is in response to Arnim's comment:
I can not leave a field without the tool complaining: "Please enter a value for this required field."
Principally this is OK, but if you hit somewhere in the middle of the table (maybe because you tried to select the tool, or just because of boredom/silliness ) you can't move away before you filled something (valid) in this cell. Would be nice to have the choice: go back and fill in vs. leave blank and go.

git-svn-id: https://subversion.int.janelia.org/ScientificComputing/Projects/IMSImageRenamer/Trunk@446 86afb187-6117-0410-83e3-87b601379a31